requirements.txt
ultralytics
opencv-python
Pillow
gTTS
pygame